requirements.txt
python-dotenv==1.0.0
python-decouple==3.8
Django==4.1
djangorestframework==3.13.1
psycopg2-binary==2.9.6
django-cors-headers==3.13.0
django-import-export==3.2.0
Pillow==9.0.1
requests==2.25.1
google-auth==2.19.1
django-ckeditor==6.7.0
